requirements.txt
brotli==1.1.0
email-validator==2.1.1
gunicorn==21.2.0
hiredis==2.3.2
pendulum==3.0.0
pillow==10.2.0
psycopg[binary,pool]==3.1.18
randomcolor==0.4.4.6
redis==5.0.2
requests-oauthlib==1.3.1
requests==2.31.0
whitenoise==6.6.0

django==5.0.2
django-allauth==0.61.1
django-anymail==10.2
django-bootstrap5==23.4
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-htmx==1.17.3
django-simple-menu==2.1.3
django-tables2==2.7.0
